+++ stac_fastapi/pgstac/core.py
@@ -9,8 +9,7 @@ import attr
 import orjson
 from asyncpg.exceptions import InvalidDatetimeFormatError
 from buildpg import render
-from fastapi import HTTPException, Request
-from pydantic import ValidationError
+from fastapi import Request
 from pygeofilter.backends.cql2_json import to_cql2
 from pygeofilter.parsers.cql2_text import parse as parse_cql2_text
 from pypgstac.hydration import hydrate
@@ -401,7 +400,14 @@ class CoreCrudClient(AsyncBaseCoreClient):
         ).get_links(extra_links=item_collection["links"])
         item_collection["links"] = links
 
-        return item_collection
+        # If we have the `fields` extension enabled
+        # we need to avoid Pydantic validation because the
+        # Items might not be a valid STAC Item objects
+        if fields := getattr(search_request, "fields", None):
+            if fields.include or fields.exclude:
+                return JSONResponse(item_collection)  # type: ignore
+
+        return ItemCollection(**item_collection)
 
     async def get_item(
         self, item_id: str, collection_id: str, request: Request, **kwargs
@@ -500,15 +506,17 @@ class CoreCrudClient(AsyncBaseCoreClient):
             filter_lang=filter_lang,
         )
 
-        # Do the request
-        try:
-            search_request = self.pgstac_search_model(**clean)
-        except ValidationError as e:
-            raise HTTPException(
-                status_code=400, detail=f"Invalid parameters provided {e}"
-            ) from e
+        search_request = self.pgstac_search_model(**clean)
+        item_collection = await self._search_base(search_request, request=request)
 
-        return await self.post_search(search_request, request=request)
+        # If we have the `fields` extension enabled
+        # we need to avoid Pydantic validation because the
+        # Items might not be a valid STAC Item objects
+        if fields := getattr(search_request, "fields", None):
+            if fields.include or fields.exclude:
+                return JSONResponse(item_collection)  # type: ignore
+
+        return ItemCollection(**item_collection)
 
     def _clean_search_args(  # noqa: C901
         self,